# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RedirectScheme": {
        "description": "Scheme applied to a redirection target (``auto`` keeps the request's).",
        "enum": [
          "auto",
          "http",
          "https"
        ],
        "title": "RedirectScheme",
        "type": "string"
      },
      "RedirectionHostCreate": {
        "description": "Payload to create a redirection host.",
        "properties": {
          "advanced_config": {
            "default": "",
            "description": "Raw nginx directives injected into the server block",
            "title": "Advanced Config",
            "type": "string"
          },
          "block_exploits": {
            "default": false,
            "description": "Block common exploit probes",
            "title": "Block Exploits",
            "type": "boolean"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Certificate for TLS termination; null serves plain :80",
            "title": "Certificate Id"
          },
          "domain_names": {
            "description": "Domains this host answers for (server_name)",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Domain Names",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "description": "Disabled hosts are excluded from config",
            "title": "Enabled",
            "type": "boolean"
          },
          "forward_domain_name": {
            "description": "Domain requests are redirected to",
            "maxLength": 255,
            "minLength": 1,
            "title": "Forward Domain Name",
            "type": "string"
          },
          "forward_http_code": {
            "default": 302,
            "description": "HTTP redirect status code (300–308)",
            "maximum": 308.0,
            "minimum": 300.0,
            "title": "Forward Http Code",
            "type": "integer"
          },
          "forward_scheme": {
            "$ref": "#/components/schemas/RedirectScheme",
            "default": "auto",
            "description": "Target scheme; 'auto' keeps the incoming request's scheme"
          },
          "hsts_enabled": {
            "default": false,
            "description": "Emit a Strict-Transport-Security header",
            "title": "Hsts Enabled",
            "type": "boolean"
          },
          "hsts_subdomains": {
            "default": false,
            "description": "Include subdomains in HSTS",
            "title": "Hsts Subdomains",
            "type": "boolean"
          },
          "http2_support": {
            "default": false,
            "description": "Enable HTTP/2 on the TLS listener",
            "title": "Http2 Support",
            "type": "boolean"
          },
          "preserve_path": {
            "default": true,
            "description": "Append the original request URI to the target",
            "title": "Preserve Path",
            "type": "boolean"
          },
          "ssl_forced": {
            "default": false,
            "description": "Redirect :80 to HTTPS",
            "title": "Ssl Forced",
            "type": "boolean"
          }
        },
        "required": [
          "domain_names",
          "forward_domain_name"
        ],
        "title": "RedirectionHostCreate",
        "type": "object"
      },
      "RedirectionHostRead": {
        "description": "Public representation of a redirection host.",
        "properties": {
          "advanced_config": {
            "default": "",
            "description": "Raw nginx directives injected into the server block",
            "title": "Advanced Config",
            "type": "string"
          },
          "block_exploits": {
            "default": false,
            "description": "Block common exploit probes",
            "title": "Block Exploits",
            "type": "boolean"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Certificate for TLS termination; null serves plain :80",
            "title": "Certificate Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "domain_names": {
            "description": "Domains this host answers for (server_name)",
            "items": {
              "type": "string"
            },
            "minItems": 1,
            "title": "Domain Names",
            "type": "array"
          },
          "enabled": {
            "default": true,
            "description": "Disabled hosts are excluded from config",
            "title": "Enabled",
            "type": "boolean"
          },
          "forward_domain_name": {
            "description": "Domain requests are redirected to",
            "maxLength": 255,
            "minLength": 1,
            "title": "Forward Domain Name",
            "type": "string"
          },
          "forward_http_code": {
            "default": 302,
            "description": "HTTP redirect status code (300–308)",
            "maximum": 308.0,
            "minimum": 300.0,
            "title": "Forward Http Code",
            "type": "integer"
          },
          "forward_scheme": {
            "$ref": "#/components/schemas/RedirectScheme",
            "default": "auto",
            "description": "Target scheme; 'auto' keeps the incoming request's scheme"
          },
          "hsts_enabled": {
            "default": false,
            "description": "Emit a Strict-Transport-Security header",
            "title": "Hsts Enabled",
            "type": "boolean"
          },
          "hsts_subdomains": {
            "default": false,
            "description": "Include subdomains in HSTS",
            "title": "Hsts Subdomains",
            "type": "boolean"
          },
          "http2_support": {
            "default": false,
            "description": "Enable HTTP/2 on the TLS listener",
            "title": "Http2 Support",
            "type": "boolean"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "preserve_path": {
            "default": true,
            "description": "Append the original request URI to the target",
            "title": "Preserve Path",
            "type": "boolean"
          },
          "ssl_forced": {
            "default": false,
            "description": "Redirect :80 to HTTPS",
            "title": "Ssl Forced",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "domain_names",
          "forward_domain_name",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "RedirectionHostRead",
        "type": "object"
      },
      "RedirectionHostUpdate": {
        "description": "Partial update of a redirection host; every field is optional.",
        "properties": {
          "advanced_config": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Advanced Config"
          },
          "block_exploits": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Block Exploits"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificate Id"
          },
          "domain_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "minItems": 1,
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Domain Names"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "forward_domain_name": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Domain Name"
          },
          "forward_http_code": {
            "anyOf": [
              {
                "maximum": 308.0,
                "minimum": 300.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forward Http Code"
          },
          "forward_scheme": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RedirectScheme"
              },
              {
                "type": "null"
              }
            ]
          },
          "hsts_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hsts Enabled"
          },
          "hsts_subdomains": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hsts Subdomains"
          },
          "http2_support": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Http2 Support"
          },
          "preserve_path": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preserve Path"
          },
          "ssl_forced": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Forced"
          }
        },
        "title": "RedirectionHostUpdate",
        "type": "object"
      },
...
      "StreamRead": {
        "description": "Public representation of a stream.",
        "properties": {
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "Certificate to terminate TLS on the TCP listener; null = plain",
            "title": "Certificate Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "enabled": {
            "default": true,
            "description": "Disabled streams are excluded from config",
            "title": "Enabled",
            "type": "boolean"
          },
          "forward_host": {
            "description": "Backend host traffic is forwarded to",
            "maxLength": 255,
            "minLength": 1,
            "title": "Forward Host",
            "type": "string"
          },
          "forward_port": {
            "description": "Backend port traffic is forwarded to",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Forward Port",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "incoming_port": {
            "description": "Port nginx listens on for this stream",
            "maximum": 65535.0,
            "minimum": 1.0,
            "title": "Incoming Port",
            "type": "integer"
          },
          "tcp_forwarding": {
            "default": true,
            "description": "Forward TCP on the incoming port",
            "title": "Tcp Forwarding",
            "type": "boolean"
          },
          "udp_forwarding": {
            "default": false,
            "description": "Forward UDP on the incoming port",
            "title": "Udp Forwarding",
            "type": "boolean"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "incoming_port",
          "forward_host",
          "forward_port",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "StreamRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "MegooPM",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/redirection-hosts": {
      "get": {
        "description": "List all redirection hosts. Admin-only.",
        "operationId": "list_redirection_hosts_api_v1_redirection_hosts_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RedirectionHostRead"
                  },
                  "title": "Response List Redirection Hosts Api V1 Redirection Hosts Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Redirection Hosts",
        "tags": [
          "redirection-hosts"
        ]
      },
      "post": {
        "description": "Create a redirection host. Admin-only.",
        "operationId": "create_redirection_host_api_v1_redirection_hosts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectionHostCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectionHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Redirection Host",
        "tags": [
          "redirection-hosts"
        ]
      }
    },
    "/api/v1/redirection-hosts/{host_id}": {
      "delete": {
        "description": "Delete a redirection host. Admin-only.",
        "operationId": "delete_redirection_host_api_v1_redirection_hosts__host_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Redirection Host",
        "tags": [
          "redirection-hosts"
        ]
      },
      "get": {
        "description": "Fetch a single redirection host. Admin-only.",
        "operationId": "get_redirection_host_api_v1_redirection_hosts__host_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectionHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Redirection Host",
        "tags": [
          "redirection-hosts"
        ]
      },
      "patch": {
        "description": "Update a redirection host. Admin-only.",
        "operationId": "update_redirection_host_api_v1_redirection_hosts__host_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "host_id",
            "required": true,
            "schema": {
              "title": "Host Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedirectionHostUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RedirectionHostRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Redirection Host",
        "tags": [
          "redirection-hosts"
        ]
      }
    },
    "/api/v1/streams": {
      "get": {
        "description": "List all streams. Admin-only.",
        "operationId": "list_streams_api_v1_streams_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/StreamRead"
                  },
                  "title": "Response List Streams Api V1 Streams Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "List Streams",
        "tags": [
          "streams"
        ]
      },
      "post": {
        "description": "Create a TCP/UDP stream forward. Admin-only.",
        "operationId": "create_stream_api_v1_streams_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Stream",
        "tags": [
          "streams"
        ]
      }
    },
    "/api/v1/streams/{stream_id}": {
      "delete": {
        "description": "Delete a stream. Admin-only.",
        "operationId": "delete_stream_api_v1_streams__stream_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "stream_id",
            "required": true,
            "schema": {
              "title": "Stream Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Stream",
        "tags": [
          "streams"
        ]
      },
      "get": {
        "description": "Fetch a single stream. Admin-only.",
        "operationId": "get_stream_api_v1_streams__stream_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "stream_id",
            "required": true,
            "schema": {
              "title": "Stream Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Stream",
        "tags": [
          "streams"
        ]
      },
      "patch": {
        "description": "Update a stream. Admin-only.",
        "operationId": "update_stream_api_v1_streams__stream_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "stream_id",
            "required": true,
            "schema": {
              "title": "Stream Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StreamUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StreamRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Stream",
        "tags": [
          "streams"
        ]
      }
    },
...
  }
}